[clang][NFC] Non-trivial fixes in C++ DR tests
