+++ clang/test/CXX/drs/cwg1xx.cpp
@@ -213,8 +213,7 @@ namespace cwg115 { // cwg115: 3.0
     &f;
     // expected-error@-1 {{reference to overloaded function could not be resolved; did you mean to call it?}}
     //   expected-note@#cwg115-f {{possible target for call}}
-    &f<int>;
-    // expected-warning@-1 {{expression result unused}}
+    (void)&f<int>;
     &g<int>;
     // expected-error@-1 {{reference to overloaded function could not be resolved; did you mean to call it?}}
     //   expected-note@#cwg115-g-int {{possible target for call}}
@@ -242,8 +241,7 @@ namespace cwg115 { // cwg115: 3.0
 
     &s.f;
     // expected-error@-1 {{cannot create a non-constant pointer to member function}}
-    &s.f<int>;
-    // expected-warning@-1 {{expression result unused}}
+    (void)&s.f<int>;
     &s.g<int>;
     // expected-error@-1 {{cannot create a non-constant pointer to member function}}
   }
@@ -269,8 +267,7 @@ namespace cwg115 { // cwg115: 3.0
 
     &s.f;
     // expected-error@-1 {{cannot create a non-constant pointer to member function}}
-    &s.f<int>;
-    // expected-warning@-1 {{expression result unused}}
+    (void)&s.f<int>;
     &s.g<int>;
     // expected-error@-1 {{cannot create a non-constant pointer to member function}}
   }
@@ -290,8 +287,7 @@ namespace cwg115 { // cwg115: 3.0
     &with_default;
     // since-cxx11-error@-1 {{reference to overloaded function could not be resolved; did you mean to call it?}}
     //   since-cxx11-note@#cwg115-with-default {{possible target for call}}
-    &with_default<>;
-    // since-cxx11-warning@-1 {{expression result unused}}
+    (void)&with_default<>;
   }
 #endif
 } // namespace cwg115
@@ -467,15 +463,17 @@ namespace cwg127 { // cwg127: 2.9
   template<typename T> struct A {
     A() { throw 0; }
     void *operator new(size_t, const char * = 0);
-    void operator delete(void *, const char *) { T::error; } // #cwg127-delete-const-char
-    // expected-error@#cwg127-delete-const-char {{type 'void' cannot be used prior to '::' because it has no members}}
-    //   expected-note@#cwg127-p {{in instantiation of member function 'cwg127::A<void>::operator delete' requested here}}
-    // expected-error@#cwg127-delete-const-char {{type 'int' cannot be used prior to '::' because it has no members}}
-    //   expected-note@#cwg127-q {{in instantiation of member function 'cwg127::A<int>::operator delete' requested here}}
-    void operator delete(void *) { T::error; }
+    void operator delete(void *, const char *) = delete; // #cwg127-delete-char
+    // cxx98-error@-1 {{deleted function definitions are a C++11 extension}}
+    void operator delete(void *) = delete;
+    // cxx98-error@-1 {{deleted function definitions are a C++11 extension}}
   };
-  A<void> *p = new A<void>; // #cwg127-p
-  A<int> *q = new ("") A<int>; // #cwg127-q
+  A<void> *p = new A<void>;
+  // expected-error@-1 {{attempt to use a deleted function}}
+  //   expected-note@#cwg127-delete-char {{'operator delete' has been explicitly marked deleted here}}
+  A<int> *q = new ("") A<int>;
+  // expected-error@-1 {{attempt to use a deleted function}}
+  //   expected-note@#cwg127-delete-char {{'operator delete' has been explicitly marked deleted here}}
 } // namespace cwg127
 
 namespace cwg128 { // cwg128: 2.7
@@ -706,7 +704,8 @@ namespace cwg141 { // cwg141: 3.1
     // FIXME: we issue a useful diagnostic first, then some bogus ones.
     b.f<int>();
     // expected-error@-1 {{no member named 'f' in 'cwg141::B'}}
-    // expected-error@-2 +{{}}
+    // expected-error-re@-2 {{{{.*}}}}
+    // expected-error-re@-3 {{{{.*}}}}
     (void)b.S<int>::n;
   }
   template<typename T> struct C {
@@ -1253,9 +1252,11 @@ namespace cwg180 { // cwg180: 2.8
 namespace cwg181 { // cwg181: 2.7
   namespace X {
     template <template X<class T> > struct A { };
-    // expected-error@-1 +{{}}
+    // expected-error@-1 {{expected '<' after 'template'}}
+    // expected-error@-2 {{expected unqualified-id}}
     template <template X<class T> > void f(A<X>) { }
-    // expected-error@-1 +{{}}
+    // expected-error@-1 {{expected '<' after 'template'}}
+    // expected-error@-2 {{expected unqualified-id}}
   }
 
   namespace Y {
@@ -1426,7 +1427,7 @@ namespace cwg197 { // cwg197: 2.7
     char &a = f(1);
     char &b = f(T(1));
     // expected-error@-1 {{non-const lvalue reference to type 'char' cannot bind to a value of unrelated type 'int'}}
-    //   expected-note@#cwg197-g-e-call {{in instantiation of function template specialization 'cwg197::g<cwg197::E>' requested here}}
+    //   expected-note@#cwg197-g-E {{in instantiation of function template specialization 'cwg197::g<cwg197::E>' requested here}}
     char &c = f(t);
     // expected-error@-1 {{non-const lvalue reference to type 'char' cannot bind to a value of unrelated type 'int'}}
   }
@@ -1436,11 +1437,9 @@ namespace cwg197 { // cwg197: 2.7
   enum E { e };
   int &f(E);
 
-  void h() {
-    g('a');
-    g(2);
-    g(e); // #cwg197-g-e-call
-  }
+template void g<char>(char);
+template void g<int>(int);
+template void g<E>(E); // #cwg197-g-E
 } // namespace cwg197
 
 namespace cwg198 { // cwg198: 2.9
