+++ clang/test/CXX/drs/cwg7xx.cpp
@@ -35,8 +35,6 @@ namespace cwg712 { // cwg712: partial
         use(a);
         use((a));
         use(cond ? a : a);
-        use((cond, a));
-        // expected-warning@-1 {{left operand of comma operator has no effect}}
 
         (void)a;
         // expected-error@-1 {{reference to local variable 'a' declared in enclosing function 'cwg712::f'}} FIXME
@@ -49,10 +47,13 @@ namespace cwg712 { // cwg712: partial
         //   expected-note@#cwg712-f-a {{'a' declared here}}
         // expected-error@#cwg712-ternary {{reference to local variable 'a' declared in enclosing function 'cwg712::f'}} FIXME
         //   expected-note@#cwg712-f-a {{'a' declared here}}
-        (void)(cond, a); // #cwg712-comma
+        (void)(cond, a);
         // expected-error@-1 {{reference to local variable 'a' declared in enclosing function 'cwg712::f'}} FIXME
         //   expected-note@#cwg712-f-a {{'a' declared here}}
-        // expected-warning@#cwg712-comma {{left operand of comma operator has no effect}}
+        // expected-warning@-3 {{left operand of comma operator has no effect}}
+
+        use((cond, a));
+        // expected-warning@-1 {{left operand of comma operator has no effect}}
       }
     };
   }
@@ -136,8 +137,8 @@ namespace cwg727 { // cwg727: partial
       // expected-error@-2 {{class template specialization of 'C' not in class 'A' or an enclosing namespace}}
       //   expected-note@#cwg727-C {{explicitly specialized declaration is here}}
       template<> void A::f<double>();
-      // expected-error@-1 {{o function template matches function template specialization 'f'}}
-      // expected-error@-2 {{non-friend class member 'f' cannot have a qualified name}}
+      // expected-error@-1 {{non-friend class member 'f' cannot have a qualified name}}
+      // expected-error@-2 {{no function template matches function template specialization 'f'}}
       template<> int A::N<double>;
       // expected-error@-1 {{non-friend class member 'N' cannot have a qualified name}}
       // expected-error@-2 {{variable template specialization of 'N' not in class 'A' or an enclosing namespace}}
@@ -188,37 +189,46 @@ namespace cwg727 { // cwg727: partial
     template<typename T> void f() { T::error; }
     // expected-error@-1 {{type 'float' cannot be used prior to '::' because it has no members}}
     //   expected-note@#cwg727-f-float {{in instantiation of function template specialization 'cwg727::D<int>::f<float>' requested here}}
-    template<typename T> static const int N = T::error;
-    // cxx98-11-error@-1 {{variable templates are a C++14 extension}}
-    // expected-error@-2 {{type 'float' cannot be used prior to '::' because it has no members}}
-    //   expected-note@#cwg727-N-float {{in instantiation of static data member 'cwg727::D<int>::N<float>' requested here}}
 
     template<> struct C<int> {};
     template<> void f<int>() {}
-    template<> const int N<int>;
 
     template<typename T> struct C<T*> {};
-    template<typename T> static const int N<T*>;
-
-    template<typename>
-    struct E {
-      template<> void f<void>() {}
-      // expected-error@-1 {{no candidate function template was found for dependent member function template specialization}}
-    };
   };
 
   void d(D<int> di) {
     D<int>::C<int>();
     di.f<int>();
-    int a = D<int>::N<int>;
-
     D<int>::C<int*>();
-    int b = D<int>::N<int*>;
-
-    D<int>::C<float>(); // #cwg727-C-float
-    di.f<float>(); // #cwg727-f-float
-    int c = D<int>::N<float>; // #cwg727-N-float
+    D<int>::C<float>();  // #cwg727-C-float
   }
+
+  template void D<int>::f<float>();  // #cwg727-f-float
+
+#if __cplusplus >= 201402L
+  template<typename>
+  struct E {
+    template<typename T> static const int N = T::error;
+    // since-cxx14-error@-1 {{type 'float' cannot be used prior to '::' because it has no members}}
+    //   since-cxx14-note@#cwg727-N-float {{in instantiation of static data member 'cwg727::E<int>::N<float>' requested here}}
+    template<> const int N<int>;
+    template<typename T> static const int N<T*>;
+  };
+  void e() {
+    int a = E<int>::N<int>;
+    int b = E<int>::N<int*>;
+    int c = E<int>::N<float>; // #cwg727-N-float
+  }
+#endif
+
+  template<typename>
+  struct F {
+    template<typename>
+    struct F2 {
+      template<> void f<void>() {}
+      // expected-error@-1 {{no candidate function template was found for dependent member function template specialization}}
+    };
+  };
 
   namespace mixed_inner_outer_specialization {
 #if __cplusplus >= 201103L
@@ -294,23 +304,21 @@ namespace cwg727 { // cwg727: partial
     template<> int f2<T>() {}
     template<> int f2<U>() {}
 
+#if __cplusplus >= 201402L
     template<typename> static int v1;
-    // cxx98-11-error@-1 {{variable templates are a C++14 extension}}
     template<> int v1<T>; // #cwg727-v1-T
     template<> int v1<U>;
-    // expected-error@-1 {{duplicate member 'v1'}}
-    //   expected-note@#cwg727-Collision-int-int {{in instantiation of template class 'cwg727::Collision<int, int>' requested here}}
-    //   expected-note@#cwg727-v1-T {{previous}}
-
+    // since-cxx14-error@-1 {{duplicate member 'v1'}}
+    //   since-cxx14-note@#cwg727-Collision-int-int {{in instantiation of template class 'cwg727::Collision<int, int>' requested here}}
+    //   since-cxx14-note@#cwg727-v1-T {{previous declaration is here}}
+#endif
+#if __cplusplus >= 201703L
     template<typename> static inline int v2;
-    // cxx98-11-error@-1 {{variable templates are a C++14 extension}}
-    // cxx98-14-error@-2 {{inline variables are a C++17 extension}}
     template<> inline int v2<T>; // #cwg727-v2-T
-    // cxx98-14-error@-1 {{inline variables are a C++17 extension}}
     template<> inline int v2<U>;
-    // cxx98-14-error@-1 {{inline variables are a C++17 extension}}
-    // expected-error@-2 {{duplicate member 'v2'}}
-    //   expected-note@#cwg727-v2-T {{previous declaration is here}}
+    // since-cxx17-error@-1 {{duplicate member 'v2'}}
+    //   since-cxx17-note@#cwg727-v2-T {{previous declaration is here}}
+#endif
 
     // FIXME: Missing diagnostic for duplicate class explicit specialization.
     template<typename> struct S1;
@@ -321,7 +329,8 @@ namespace cwg727 { // cwg727: partial
     template<> struct S2<T> {}; // #cwg727-S2-T
     template<> struct S2<U> {};
     // expected-error@-1 {{redefinition of 'S2<int>'}}
-    //   expected-note@#cwg727-S2-T {{previous}}
+    //   cxx98-11-note@#cwg727-Collision-int-int {{in instantiation of template class 'cwg727::Collision<int, int>' requested here}}
+    //   expected-note@#cwg727-S2-T {{previous definition is here}}
   };
   Collision<int, int> c; // #cwg727-Collision-int-int
 } // namespace cwg727
