+++ clang/test/CXX/drs/cwg9xx.cpp
@@ -59,29 +59,33 @@ struct A {
 struct B : private A { // #cwg952-B
 };
 struct C : B {
-  void f() {
-    I i1;
-    // expected-error@-1 {{'I' is a private member of 'cwg952::example1::A'}}
-    //   expected-note@#cwg952-B {{constrained by private inheritance here}}
-    //   expected-note@#cwg952-I {{member is declared here}}
-  }
-  I i2;
+  void f();
+  I i1;
   // expected-error@-1 {{'I' is a private member of 'cwg952::example1::A'}}
   //   expected-note@#cwg952-B {{constrained by private inheritance here}}
   //   expected-note@#cwg952-I {{member is declared here}}
   struct D {
-    I i3;
+    I i2;
     // expected-error@-1 {{'I' is a private member of 'cwg952::example1::A'}}
     //   expected-note@#cwg952-B {{constrained by private inheritance here}}
     //   expected-note@#cwg952-I {{member is declared here}}
-    void g() {
-      I i4;
-      // expected-error@-1 {{'I' is a private member of 'cwg952::example1::A'}}
-      //   expected-note@#cwg952-B {{constrained by private inheritance here}}
-      //   expected-note@#cwg952-I {{member is declared here}}
-    }
+    void g();
   };
 };
+
+void C::f() {
+  I i3;
+  // expected-error@-1 {{'I' is a private member of 'cwg952::example1::A'}}
+  //   expected-note@#cwg952-B {{constrained by private inheritance here}}
+  //   expected-note@#cwg952-I {{member is declared here}}
+}
+
+void C::D::g() {
+  I i4;
+  // expected-error@-1 {{'I' is a private member of 'cwg952::example1::A'}}
+  //   expected-note@#cwg952-B {{constrained by private inheritance here}}
+  //   expected-note@#cwg952-I {{member is declared here}}
+}
 } // namespace example1
 namespace example2 {
 struct A {
